sqlean: Enhance sqlean_version() to check features
This was the simplest way I saw to check feature availability at
runtime, but I'm not sure if the interface is the most sensible. The
value returned from sqlean_version('SQLEAN_VSV') for example is not a
version number, but a truth / boolean value.

Perhaps sqlean_check($arg) or just sqlean($arg) reads better?

    sqlite> SELECT sqlean_check('HAVE_TIMESPEC_GET');
    sqlite> SELECT sqlean_check('SQLEAN_IPADDR');

    sqlite> SELECT sqlean('HAVE_ARPA_INET_H');
    sqlite> SELECT sqlean('SQLEAN_UNICODE');