sqlean: New C function sqlean_unsupported()
sqlean_ununsupported() is a "filler" for extensions to use when a
particular feature is unavailable. It is registered in place of the
non-working function which should be compliled out. In the event that
it is called by user code it raises a run-time error.

The intended use-case is mostly for test files: by allowing unsupported
functions to still be prepared, we can use the same unmodified test
files everywhere regardless feature availablity.