build: New script for running SQL tests
test/.run is a test runner for the new build system; it is not a
user-facing tool. Developer, end-user or packaging testing should use
"make test" or "make test-$EXT".

This makes testing more robust and informative, but I consider it only
half an improvement. It would be better to have a proper testing
framework in SQL that doesn't rely on parsing output, and gives better
feedback on error details. Ideally something based on the Test Anything
Protocol (TAP).